Make the following changes:
,

Find this element:

  <CustomProfileChart pointAName="Site A" pointBName="Site B" isStale=true totalDistanceKm=0.7 isActionPending=false>

Relevant files:

    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_8c671237._.js
    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_8c671237._.js
    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_8c671237._.js
    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

now tower height is seen when we adjust it. now we want to show red dot on our los line whereever it is being blocked by elevation ground. so that we can see all the points on line where our line is touching the ground. also both of our tower are on the edge. we need to set ratio. like we need whole link to use 80% part of elevation profile so that each tower has 10% space left. and we need that we can also change the tower location by moving the line same way we are adjusting its height.