Make the following changes:
,

Find this element:

  <CustomProfileChart pointAName="Site A" pointBName="Site B" isStale=true totalDistanceKm=0.7 isActionPending=false>

Relevant files:

    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_8c671237._.js
    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_8c671237._.js
    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_8c671237._.js
    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

still the ratio is not set. we need to fetch elevation profile whole link but should have 20 % more so that tower back area is also visible. 2. we should be able to move tower on elevation terrain so that we can check realtime how our line is moving and whenever it touches the elevation then on that part of line red dot should appear showing us impace zome. 3. we need to change tower height not in decimals. just in numbers. 4. whenever user leaves the tower then automatically take that tower height and complete the analysis so that there is need of re-analyse button. also for moving the line and adjusting tower height dont require to recall elevation api as if it is already called once for same location then we have data and that can be used. 5. we need to have aerial distance always shown above the line connecting two towers. when we are adjusting tower by moving tower point then it should also change realtime according to distance. so that we can keep track of aerial distance. also when we move tower then automatically re-analysis must be done. so that there is no need to do it again and again.