Make the following changes:
,

Find this element:

  <main class="flex-1 overflow-y-auto">

with the text content:

    ←Move left→Move right↑Move up↓Move down+Zoom in-Zoom outHomeJump left by 75%EndJump right by 75%Page UpJump up by 75%Page DownJump down by 75%To navigate, press the arrow keys.MapTerrainSatelliteLabelsKeyboard shortcutsMap DataImagery ©2025 Airbus, Maxar TechnologiesImagery ©2025 Airbus, Maxar Technologies50 m Click to toggle between metric and imperial unitsTermsReport a map errorStart Link Analysis Site ANameLatitudeLongitudeTower Height (m)26m  NEEDS RE-ANALYZEAerial Dist.700mMin. Clear.20.0mReq. Fresnel (m):Re-Analyze Site BNameLatitudeLongitudeTower Height (m)20m

Relevant files:

And change it as follows:

map was earlier showing full. and when we click start link analysis button then only bottom panel shows above the map. it must look as if its hovering above the map. its feel should be like it as appeared from expanding of start link analysis button and when we want it to hide then it again wraps back into button animation. so we always need full map for whole website. and bottom panel must be translusent to that behind map can also be seen slightly