Make the following changes:
,

Find this element:

  <CustomProfileChart pointAName="Site A" pointBName="Site B" isStale=true totalDistanceKm=0.7 isActionPending=false>

Relevant files:

    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_8c671237._.js
    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_8c671237._.js
    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_8c671237._.js
    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1747861782075.cluster-sumfw3zmzzhzkx4mpvz3ogth4y.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

we need elevation profile to use most space from bottom to left and right. so that it fills most part and looks elegant and aesthetic for website. understand how its colours must be so that it looks nice and bright. we also need distance to be shown below the elevation profile. currently its showing starting distance at left most part of elevation profile and total distance at bottom rightmost part of the chart. we need it show at different intervals.