we want to make this responsive for mobile. as our main users would be mobile users. So we need to take care of touch things. we need website logo to be removed for right now. also we need to have a make a report button. this would be the most important feature of our website. which will make downloadable word file out of the checked link. it must have Heading as: LOS Survey Report: "Point A" name to "Point B" name
then it would have details in Table format as
point a, lat of point a, long of point a, point b, lat of point b, long of point b, aerial distance, fresenl height considered, tower at point a, tower at point b, los possible/not

then it will have remarks:

As per desktop survey, "LOS is Possible" from "Point A" to "Point B". Aerial distance between two sites is "Aerial Distance in KM or m". We have considered tower height as __ for point a and ___ for point b and fresnel height as ___ for this survey. Physical survey is recommended for furthur analysis of the site.

then it will have snapshot of map where both the markers and line with aerial distance written would be seen

and at last it will have snapshot of our elevation profile where tower height and los line is seen connecting two towers with aerial distance written above line.