<BottomPanel>
we need this panel to be of fixed according to the ratio. we  dont want to scroll it up and down. we need its height to be fixed as per ratio of map seen and bottom panel seen. we need height fixed. we need footer to be seen all the time. from footer remove all rights reserve by nav wireless line.

also when screen is too small like in mobile we need it to scroll sideways from left to right and vise versa.

also make LiFi link Pro branding more. use Blue color as linkedin ones. and set word spacing little more in heading name. its our application's name. so that will make our website look more cool. Also we need LiFi Link Pro name to be seen in Elevation profile as a background like watermark in middle inside terrain. it looks cool. Also the button of hiding the panel down and bringing it up must be translucent.