here we are having some errors when making new links. i want to update logic of creating new links on map again. when first time website is opening we need to only show "Add new link" button on the home page. which is now been shown correctly. but there is again add new link button which would be duplication if shown two times on the screen. when we click Add New link button then currently it is opening <bottompanel>. in there we need <card> title as Point A and Point B for both the cards. In name we should have prefilled name as Site A for Point A and Site B for Point B. We need coordinates as merged of latitude and longitude both inputs. we need to accept all sorts of available formats of coordinates like dms and decimal in all formats. for example 1. lat, long 2. lat,long 3. lat long 4. (lat,long) 5. (lat, long)
all in dms as well as decimal format.

we dont want <primitive.span> in both the cards as its already doable via adjusting tower height in elevation panel. so we need to remove that.

now when we click analyze link this should give us los status and generate report button which is currently working as desired.

but after that we need our button of add another link to showup. when we click add new link then again whole process of opening bottom panel should start.

we also want to search on google map locations so that its easy for users to use map and find locations. so make map such that is easy to navigate to location.

and solve every error as possible