<BottomPanel>
Right now its not able to include all elements. we need to resize each elements so that it fits perfectly in fixed height.
- We dont need <primitive.span> from both the <card> as we dont now need slider.

also when we are using same website for mobile where might require to decrese the height,. we can decrese little bit. but we must know that we have to make sure all elements are resized according to the envirnoment they are in. website must be responsive. each element must be responsive. so that whole structure looks good to look.
- in mobile type window where width is small we can make it horizontal scrolling which slides like images cascading in slideshow we need that slide between the three panels in bottom panel.

also we need middle part where elevation profile is there, we need elevation profile to use most part of middle section. currently there is loads of blank space around the elevation profile. we also need elevation chart to extend beyond both the towers so that towers dont look on the edge of the elevation panel. we need to fetch 10% more elevation for both side of tower more to fill the extra beyond tower elevation.